[extension] create new class for the video recommendations display, use of the new class inside of TournesolRecommendations, TournesolRecommendations takes care of retrieving recommendations, TournesolComponent takes care of generating the tournesol container with his action buttons and banner, TournesolVideosBox takes care of generating video box items
